requirements.txt
PyYAML>=6.0,<7
requests>=2.31,<3
websockets>=12,<16
pandas>=2.0,<3
pytest>=8,<9
pytest-asyncio>=0.23,<1
googletrans==4.0.0rc1
